requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
# requests_html==0.10.0
requests==2.26.0
motor==2.5.1
pymongo==3.12.0
beautifulsoup4==4.10.0
selenium==4.1.0
python-dotenv==0.19.2